IoUring: Refactor IoUringIoOps to better match the underlying io_urin… (#14521)
…g_sqe

Modification:

We should be able to support any io_uring operation in the future for
this we should allow to setup the right struct.

Modifications:

- Change constructor to match underlying struct
- Remove public getters as these are only used internally and might
change in the future
- Rename static fields

Result:

More flexible implementation for io_uring

+++ transport-classes-io_uring/src/main/java/io/netty/channel/uring/AbstractIoUringChannel.java
@@ -253,7 +253,7 @@ abstract class AbstractIoUringChannel extends AbstractChannel implements UnixCha
         if (registration != null) {
             if (socket.markClosed()) {
                 int fd = fd().intValue();
-                IoUringIoOps ops = IoUringIoOps.newClose(fd, 0, nextOpsId());
+                IoUringIoOps ops = IoUringIoOps.newClose(fd, (byte) 0, nextOpsId());
                 registration.submit(ops);
             }
         } else {
@@ -345,7 +345,7 @@ abstract class AbstractIoUringChannel extends AbstractChannel implements UnixCha
         int fd = fd().intValue();
         IoUringIoRegistration registration = registration();
         IoUringIoOps ops = IoUringIoOps.newPollAdd(
-                fd, 0, Native.POLLOUT, (short) Native.POLLOUT);
+                fd, (byte) 0, Native.POLLOUT, (short) Native.POLLOUT);
         pollOutId = registration.submit(ops);
         if (pollOutId != 0) {
             ioState |= POLL_OUT_SCHEDULED;
@@ -357,7 +357,7 @@ abstract class AbstractIoUringChannel extends AbstractChannel implements UnixCha
         int fd = fd().intValue();
         IoUringIoRegistration registration = registration();
         IoUringIoOps ops = IoUringIoOps.newPollAdd(
-                fd, 0, Native.POLLRDHUP, (short) Native.POLLRDHUP);
+                fd, (byte) 0, Native.POLLRDHUP, (short) Native.POLLRDHUP);
         pollRdhupId = registration.submit(ops);
         if (pollRdhupId != 0) {
             ioState |= POLL_RDHUP_SCHEDULED;
@@ -533,7 +533,7 @@ abstract class AbstractIoUringChannel extends AbstractChannel implements UnixCha
                         int fd = fd().intValue();
                         // Best effort to cancel the already submitted connect request.
                         registration.submit(IoUringIoOps.newAsyncCancel(
-                                fd, 0, connectId, Native.IORING_OP_CONNECT));
+                                fd, (byte) 0, connectId, Native.IORING_OP_CONNECT));
                     }
                     cancelOutstandingReads(registration, numOutstandingReads);
                     cancelOutstandingWrites(registration, numOutstandingWrites);
@@ -663,7 +663,7 @@ abstract class AbstractIoUringChannel extends AbstractChannel implements UnixCha
             int fd = fd().intValue();
             IoUringIoRegistration registration = registration();
             IoUringIoOps ops = IoUringIoOps.newPollAdd(
-                    fd, 0, Native.POLLIN, (short) Native.POLLIN);
+                    fd, (byte) 0, Native.POLLIN, (short) Native.POLLIN);
             pollInId = registration.submit(ops);
             if (pollInId != 0) {
                 ioState |= POLL_IN_SCHEDULED;
@@ -982,7 +982,7 @@ abstract class AbstractIoUringChannel extends AbstractChannel implements UnixCha
 
                     int fd = fd().intValue();
                     IoUringIoRegistration registration = registration();
-                    IoUringIoOps ops = IoUringIoOps.newSendmsg(fd, 0, Native.MSG_FASTOPEN,
+                    IoUringIoOps ops = IoUringIoOps.newSendmsg(fd, (byte) 0, Native.MSG_FASTOPEN,
                             hdr.address(), hdr.idx());
                     connectId = registration.submit(ops);
                 } else {
@@ -1038,7 +1038,7 @@ abstract class AbstractIoUringChannel extends AbstractChannel implements UnixCha
         int fd = fd().intValue();
         IoUringIoRegistration registration = registration();
         IoUringIoOps ops = IoUringIoOps.newConnect(
-                fd, 0, remoteAddressMemoryAddress, nextOpsId());
+                fd, (byte) 0, remoteAddressMemoryAddress, nextOpsId());
         connectId = registration.submit(ops);
     }
 
@@ -1075,21 +1075,21 @@ abstract class AbstractIoUringChannel extends AbstractChannel implements UnixCha
         int fd = fd().intValue();
         if ((ioState & POLL_RDHUP_SCHEDULED) != 0) {
             registration.submit(IoUringIoOps.newPollRemove(
-                    fd, 0, pollRdhupId, (short) Native.POLLRDHUP));
+                    fd, (byte) 0, pollRdhupId, (short) Native.POLLRDHUP));
         }
         if ((ioState & POLL_IN_SCHEDULED) != 0) {
             registration.submit(IoUringIoOps.newPollRemove(
-                    fd, 0, pollInId, (short) Native.POLLIN));
+                    fd, (byte) 0, pollInId, (short) Native.POLLIN));
         }
         if ((ioState & POLL_OUT_SCHEDULED) != 0) {
             registration.submit(IoUringIoOps.newPollRemove(
-                    fd,  0, pollOutId, (short) Native.POLLOUT));
+                    fd,  (byte) 0, pollOutId, (short) Native.POLLOUT));
         }
 
         if (AbstractIoUringChannel.this.connectPromise != null && connectId != 0) {
             // Best effort to cancel the already submitted connect request.
             registration.submit(IoUringIoOps.newAsyncCancel(
-                    fd, 0, connectId, Native.IORING_OP_CONNECT));
+                    fd, (byte) 0, connectId, Native.IORING_OP_CONNECT));
         }
         cancelOutstandingReads(registration, numOutstandingReads);
         cancelOutstandingWrites(registration, numOutstandingWrites);
