+++ transport-classes-io_uring/src/main/java/io/netty/channel/uring/AbstractIoUringServerChannel.java
@@ -80,7 +80,7 @@ abstract class AbstractIoUringServerChannel extends AbstractIoUringChannel imple
             assert numOutstandingReads == 1;
             int fd = fd().intValue();
             IoUringIoOps ops = IoUringIoOps.newAsyncCancel(
-                    fd, 0, acceptId, Native.IORING_OP_ACCEPT);
+                    fd, (byte) 0, acceptId, Native.IORING_OP_ACCEPT);
             registration.submit(ops);
         } else {
             assert numOutstandingReads == 0;
@@ -120,7 +120,7 @@ abstract class AbstractIoUringServerChannel extends AbstractIoUringChannel imple
 
             int fd = fd().intValue();
             IoUringIoRegistration registration = registration();
-            IoUringIoOps ops = IoUringIoOps.newAccept(fd, 0, 0,
+            IoUringIoOps ops = IoUringIoOps.newAccept(fd, (byte) 0, 0,
                     acceptedAddressMemoryAddress, acceptedAddressLengthMemoryAddress, nextOpsId());
             acceptId = registration.submit(ops);
             return 1;
